Add files via upload
In this project, I am using a 6-axis Articulated Arm robot manufactured by Mitsubishi and the 
software is MELFA ToolBox V. The robot’s primary task is to fill the gas cylinder with a gas 
nozzle. Cylinders as inputs will be waiting in the conveyor belt, and the robot will take 
cylinders one by one from the conveyor belt to the box. The box measures the weight of gas in 
the cylinder and it also works as a carrier of gas cylinders. The carrier is a pallet box of 2 by 2.
The robot should complete all the tasks automatically without any interruption. Here
interruption means an emergency or defect in a cylinder. If there is any defect in the cylinder 
the weight sensor can detect it and it will send the signal to the operator.